feat: migrate report sidebar to in global studio sidebar
